+++ js/main.js
@@ -1,91 +1,6 @@
-const PICTURE_COUNT = 25;
-const AVATAR_COUNT = 6;
-const LIKE_MIN_COUNT = 15;
-const LIKE_MAX_COUNT = 200;
-const COMMENT_COUNT = 10;
-const COMMENT_LINES = [
-  'Всё отлично!',
-  'В целом всё неплохо.Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент ?!'
-];
+import './util.js';
+import { getPictures } from './data.js';
 
-const DESCRIPTIONS = [
-  'Летний чил на югаx. #тай #отдых #лето #чил #travel',
-  'Тестим новую камеру! #camera #test #new',
-  'Затусили с друзьями на моpе #laptevsea #north #northeastpassage',
-  'Как же круто тут кормят #food #foodgram #instafood',
-  'Отдыхаем… #chill #relax #group #photo',
-  'Цените каждое мгновенье.',
-  'Вот это тачка! #wow #car #carwow #drive'
-];
-
-const NAMES = [
-  'Никита',
-  'Антон',
-  'Анна',
-  'Елена',
-  'Олег',
-  'Артём',
-];
-
-// Получение случайного идентификатора
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-// Получение случайного элемента массива
-const getRandomArrayElement = (Array) =>
-  Array[getRandomInteger(0, Array.length - 1)];
-
-// Создание идентификатора с увеличением на единицу
-const createIdGenerator = () => {
-  let lastGeneratedId = 0;
-
-  return () => {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-};
-
-const generateCommentId = createIdGenerator();
-
-// Создание массива случайных сообщений
-const createMessage = () =>
-  Array.from({ length: getRandomInteger(1, 2) }, () =>
-    getRandomArrayElement(COMMENT_LINES)
-  ).join(' ');
-
-// Создание объекта с комментарием (id, аватарка, комментарий, имя комментатора)
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
-  message: createMessage(),
-  name: getRandomArrayElement(NAMES),
-});
-
-// Создание объекта (id, фото, описание, кол-во лайков, массив случайного кол-ва комментариев)
-const createPicture = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(LIKE_MIN_COUNT, LIKE_MAX_COUNT),
-  comments: Array.from(
-    { length: getRandomInteger(0, COMMENT_COUNT) },
-    createComment
-  ),
-});
-
-// Создание массива объектов
-const getPictures = () =>
-  Array.from(
-    { length: PICTURE_COUNT }, (_, pictureIndex) =>
-    createPicture(pictureIndex + 1)
-  );
-
-getPictures();
+console.log(
+  getPictures()
+);
